requirements.txt
aniso8601==7.0.0
certifi==2018.8.24
chardet==3.0.4
Click==7.0
cycler==0.10.0
ffmpeg==1.4
ffmpy==0.2.2
Flask==1.0.2
Flask-Cors==3.0.8
Flask-PyMongo==2.2.0
Flask-RESTful==0.3.7
google-api-core==1.14.3
google-auth==1.7.1
google-cloud-automl==0.9.0
google-cloud-core==1.0.3
google-cloud-storage==1.23.0
google-resumable-media==0.5.0
googleapis-common-protos==1.6.0
gunicorn==19.9.0
httplib2==0.11.3
idna==2.7
itsdangerous==0.24
Jinja2==2.10
kiwisolver==1.0.1
MarkupSafe==1.0
matplotlib==3.0.1
newrelic==4.20.1.121
numpy==1.15.4
oauth2==1.9.0.post1
pydub==0.23.1
pymongo==3.7.2
pyparsing==2.3.0
python-dateutil==2.7.5
pytz==2019.1
requests==2.19.1
six==1.12.0
sqlparse==0.3.0
urllib3==1.23
virtualenv==16.2.0
Werkzeug==0.14.1
